backend/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
googletrans==4.0.0rc1
grpcio==1.68.0
grpcio-status==1.68.0
gTTS==2.5.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.11.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.5.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.0
pydantic_core==2.27.0
pygame==2.6.1
pyparsing==3.2.0
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
sniffio==1.3.1
tqdm==4.67.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.1.3
zipp==3.21.0
